+++ packages/grid-export/src/features/grid-copy-to-clipboard/transformers/__tests__/HeaderTransformer.test.ts
@@ -28,7 +28,7 @@ describe("HeaderTransformer", () => {
   });
   describe("transformGroupHeaders", () => {
     const r = transformGroupHeaders(testGroupConfigsWithGroups);
-    expect(r[0]).toBe('<th colspan="2">Name</th>');
-    expect(r[1]).toBe('<th colspan="1">Info</th>');
+    expect(r[0]).toBe('<th style="text-align: left" colspan="2">Name</th>');
+    expect(r[1]).toBe('<th style="text-align: left" colspan="1">Info</th>');
   });
 });